+++ src/lib.rs
@@ -7,7 +7,7 @@ pub mod consts;
 pub mod defaults;
 pub mod options;
 
-use consts::RESPONSIVE_SIZES;
+use consts::{VARIANTS, VARIANT_SEARCHER};
 use defaults::{RE, SORTER};
 use options::Options;
 
@@ -52,41 +52,16 @@ fn sort_classes_vec<'a>(classes: impl Iterator<Item = &'a str>) -> Vec<&'a str> 
     for (class, maybe_size) in enumerated_classes {
         match maybe_size {
             Some(size) => tailwind_classes.push((class, size)),
-            None => match class.as_bytes() {
-                [b's', b'm', b':', ..] => responsive.entry("sm").or_insert(Vec::new()).push(class),
+            None => match VARIANT_SEARCHER.find(&class) {
+                Some(prefix_match) => {
+                    let prefix = VARIANTS[prefix_match.pattern()];
+                    responsive
+                        .entry(prefix)
+                        .or_insert_with(Vec::new)
+                        .push(class)
+                }
 
-                [b'm', b'd', b':', ..] => responsive.entry("md").or_insert(Vec::new()).push(class),
-
-                [b'l', b'g', b':', ..] => responsive.entry("lg").or_insert(Vec::new()).push(class),
-
-                [b'x', b'l', b':', ..] => responsive.entry("xl").or_insert(Vec::new()).push(class),
-
-                [b'2', b'x', b'l', b':', ..] => responsive
-                    .entry("2xl")
-                    .or_insert_with(|| vec![])
-                    .push(class),
-
-                [b'3', b'x', b'l', b':', ..] => responsive
-                    .entry("3xl")
-                    .or_insert_with(|| vec![])
-                    .push(class),
-
-                [b'4', b'x', b'l', b':', ..] => responsive
-                    .entry("4xl")
-                    .or_insert_with(|| vec![])
-                    .push(class),
-
-                [b'5', b'x', b'l', b':', ..] => responsive
-                    .entry("5xl")
-                    .or_insert_with(|| vec![])
-                    .push(class),
-
-                [b'6', b'x', b'l', b':', ..] => responsive
-                    .entry("6xl")
-                    .or_insert_with(|| vec![])
-                    .push(class),
-
-                _ => custom_classes.push(class),
+                None => custom_classes.push(class),
             },
         }
     }
@@ -100,9 +75,9 @@ fn sort_classes_vec<'a>(classes: impl Iterator<Item = &'a str>) -> Vec<&'a str> 
 
     let mut sorted_responsive_classes = vec![];
 
-    for key in RESPONSIVE_SIZES.iter() {
+    for key in VARIANTS.iter() {
         let (mut sorted_classes, new_custom_classes) = sort_responsive_classes(
-            responsive.remove(key).unwrap_or_else(|| vec![]),
+            responsive.remove(key).unwrap_or_else(Vec::new),
             custom_classes,
             key.len() + 1,
         );